refactor: rename plugin modules to not have a name conflict with the classes
when naming the files the same as the class, the module takes higher
priority over anything defined in __init__.py when importing

this naming is not PEP8 compatible
