+++ threeML/test/test_plugin_loading.py
@@ -3,7 +3,7 @@ __author__ = "drjfunk"
 import os
 
 from threeML.io.file_utils import within_directory
-from threeML.plugins.OGIPLike import OGIPLike
+from threeML.plugins import OGIPLike
 from threeML.plugins.SwiftXRTLike import SwiftXRTLike
 
 from .conftest import get_test_datasets_directory
