+++ threeML/utils/spectrum/share_spectrum.py
@@ -1,8 +1,8 @@
 import numpy as np
 
 from threeML.io.logging import setup_logger
-from threeML.plugins.DispersionSpectrumLike import DispersionSpectrumLike
-from threeML.plugins.SpectrumLike import SpectrumLike
+from threeML.plugins import DispersionSpectrumLike
+from threeML.plugins import SpectrumLike
 
 log = setup_logger(__name__)
