+++ threeML/utils/time_series/polynomial.py
@@ -26,8 +26,8 @@ from threeML.minimizer.minimization import (
     GlobalMinimization,
     LocalMinimization,
 )
-from threeML.plugins.UnbinnedPoissonLike import EventObservation, UnbinnedPoissonLike
-from threeML.plugins.XYLike import XYLike
+from threeML.plugins import EventObservation, UnbinnedPoissonLike
+from threeML.plugins import XYLike
 
 log = setup_logger(__name__)
